File needs to be a ReadStream
While using it together with Typescript, the compiler was saying to use file as File type. Although if you use it as File type you are going to have the following error:

TypeError: source.on is not a function
    at DelayedStream.create (/Users/pedrogabriel/healthmeet/node_modules/delayed-stream/lib/delayed_stream.js:33:10)
    at CombinedStream.append (/Users/pedrogabriel/healthmeet/node_modules/combined-stream/lib/combined_stream.js:45:37)
    at FormData.append (/Users/pedrogabriel/healthmeet/node_modules/form-data/lib/form_data.js:75:3)
    at Object.<anonymous> (/Users/pedrogabriel/healthmeet/node_modules/openai/dist/api.js:572:36)
    at Generator.next (<anonymous>)
    at /Users/pedrogabriel/healthmeet/node_modules/openai/dist/api.js:21:71
    at new Promise (<anonymous>)
    at __awaiter (/Users/pedrogabriel/healthmeet/node_modules/openai/dist/api.js:17:12)
    at Object.createTranscription (/Users/pedrogabriel/healthmeet/node_modules/openai/dist/api.js:555:108)
    at Object.<anonymous> (/Users/pedrogabriel/healthmeet/node_modules/openai/dist/api.js:1209:75)

Because values of formData object can only by any of these: string, Buffer, Stream. One solution is either change the type on the API contract or convert the file to Stream before posting it or change the API signature. Even at openai docs, before sending the file the method fs.createReadStream(<file>) is called.

All in all, if we want to use the file Path the fs.createReadStream should be inside Openai lib.
If we want to not change the openai lib, we need to change the signature to ReadStream

const { Configuration, OpenAIApi } = require("openai");
const configuration = new Configuration({
  apiKey: process.env.OPENAI_API_KEY,
});
const openai = new OpenAIApi(configuration);
const resp = await openai.createTranscription(
  fs.createReadStream("audio.mp3"),
  "whisper-1"
);